requirements.txt
pytest>=7.0
allure-pytest>=2.13
selenium>=4.0
webdriver-manager>=4.0